PyTorch_emnist/practice-cnn/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
torch==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tqdm
matplotlib==3.10.0
numpy==1.26.4
pillow==11.0.0
